Single folder for both versions
